feat(VideoPlayer): refactor video component to use method in hook
